# internal/services/proto/newproto.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/newproto.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "ProfileService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/profiles": {
      "get": {
        "summary": "Получение списка профилей",
        "operationId": "ProfileService_GetProfiles",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileGetProfilesResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ]
      },
      "post": {
        "summary": "Создание профиля",
        "operationId": "ProfileService_CreateProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileCreateProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/profileCreateProfileRequest"
            }
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    },
    "/v1/profiles/delete/id/{id}": {
      "delete": {
        "summary": "Удаление профиля по id",
        "operationId": "ProfileService_DeleteProfileByID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileDeleteProfileByIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    },
    "/v1/profiles/id/{id}": {
      "get": {
        "summary": "Получение профиля по id",
        "operationId": "ProfileService_GetProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileGetProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    },
    "/v1/profiles/token": {
      "get": {
        "summary": "Get profile by token",
        "description": "Retrieves a profile using the authorization token from the header.",
        "operationId": "ProfileService_GetProfileByToken",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileGetProfileByTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/v1/profiles/update": {
      "put": {
        "summary": "Обновление профиля",
        "operationId": "ProfileService_UpdateProfile",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileUpdateProfileResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/profileUpdateProfileRequest"
            }
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    },
    "/v1/profiles/userid/{userId}": {
      "get": {
        "summary": "Получение профиля по user_id",
        "operationId": "ProfileService_GetProfileByUserID",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/profileGetProfileByUserIDResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "ProfileService"
        ]
      }
    }
  },
  "definitions": {
    "profileCreateProfileRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "additionalData": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "title": "Запрос для создания профиля"
    },
    "profileCreateProfileResponse": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "title": "Ответ для создания профиля"
    },
    "profileDeleteProfileByIDResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "title": "Ответ для удаления профиля"
    },
    "profileGetProfileByTokenResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/profileProfile"
        },
        "error": {
          "type": "string"
        }
      },
      "title": "Ответ с данными профиля"
    },
    "profileGetProfileByUserIDResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/profileProfile"
        }
      },
      "title": "Ответ для получения профиля"
    },
    "profileGetProfileResponse": {
      "type": "object",
      "properties": {
        "profile": {
          "$ref": "#/definitions/profileProfile"
        }
      },
      "title": "Ответ для получения профиля"
    },
    "profileGetProfilesResponse": {
      "type": "object",
      "properties": {
        "profiles": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/profileProfile"
          },
          "title": "Список профилей"
        },
        "totalRecords": {
          "type": "integer",
          "format": "int32",
          "title": "Общее количество записей"
        }
      },
      "title": "Ответ для получения списка профилей"
    },
    "profileProfile": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "additionalData": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string"
        },
        "updatedAt": {
          "type": "string"
        }
      },
      "title": "Сообщение профиля"
    },
    "profileUpdateProfileRequest": {
      "type": "object",
      "properties": {
        "userId": {
          "type": "string"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "middleName": {
          "type": "string"
        },
        "phone": {
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "birthday": {
          "type": "string"
        },
        "additionalData": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "title": "Запрос для обновление профиля"
    },
    "profileUpdateProfileResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "message": {
          "type": "string"
        }
      },
      "title": "Ответ для обновления профиля"
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "description": "Bearer token",
      "name": "Authorization",
      "in": "header"
    }
  }
}